+++ go-app/frontend/src/swr/useControllerConfiguration.ts
@@ -5,7 +5,9 @@ import { main } from "../../wailsjs/go/models";
 export const useControllerConfiguration = (controller:  main.Interop_GenericController) => {
   return useSWR(
     ["system", "controller", controller.UniqueID, "configuration"],
-    async () => GetControllerConfiguration(controller.UniqueID),
+    async (): Promise<main.Interop_ControllerConfiguration | null> => {
+      return GetControllerConfiguration(controller.UniqueID)
+    },
     { suspense: true, revalidateOnMount: true },
   );
 };